Address comments modulo:
1. Add tests for TaskSource
2. Add docs for TaskSource
3. Get existing failing tests to pass

+++ fml/task_queue_id.h
@@ -9,13 +9,19 @@
 
 namespace fml {
 
+/**
+ * `MessageLoopTaskQueues` task dispatcher's internal task queue identifier.
+ */
 class TaskQueueId {
  public:
+  /// This constant indicates whether a task queue has been subsumed by a task
+  /// runner.
   static const size_t kUnmerged;
 
+  /// Intializes a task queue with the given value as it's ID.
   explicit TaskQueueId(size_t value) : value_(value) {}
 
-  operator int() const { return value_; }
+  operator size_t() const { return value_; }
 
  private:
   size_t value_ = kUnmerged;
